TPSAInterface

Documentation for TPSAInterface.

TPSAInterface.is_tps — Function

Returns TPSAInterface.IsTPS() iftis a TPS, elseTPSAInterface.IsNotTPS().

TPSAInterface.is_tps_type — Function

Returns TPSAInterface.IsTPSType() iftis a TPS type, elseTPSAInterface.IsNotTPSType().

TPSAInterface.numtype — Function

Returns the number type of the monomial coefficients of the TPS. t may be a TPS or TPS Type.

TPSAInterface.nvars — Function

Returns the number of variables in the TPSA. t may be a TPS or TPS Type.

TPSAInterface.nparams — Function

Returns the number of parameters in the TPSA. t may be a TPS or TPS Type.

TPSAInterface.ndiffs — Function

Returns the number of variables + parameters in the TPSA. t may be a TPS or TPS Type.

TPSAInterface.maxord — Function

Returns the maximum truncation order of the TPSA. t may be a TPS or TPS Type.

TPSAInterface.numcoefs — Function

Returns the number of monomial coefficients in the TPSA. t may be a TPS or TPS Type.

TPSAInterface.scalar — Function

Returns the zeroth order (scalar) part of the TPS t.

TPSAInterface.norm_tps — Function

Returns a norm of the passed TPS t including all monomial coefficients.

TPSAInterface.clear! — Function

Sets all monomial coefficients in t equal to 0.

TPSAInterface.geti — Function

geti(t, i::Integer)

Gets the i-th monomial coefficient in the TPS t, where i=0 specifies the scalar part, and the monomials are sorted by order.

TPSAInterface.getm — Function

getm(t, mono::AbstractArray{<:Integer})

Gets the coefficient of the monomial with orders mono.

TPSAInterface.seti! — Function

seti!(t, v, i::Integer)

Sets the i-th monomial coefficient in the TPS t to v, where i=0 specifies the scalar part, and the monomials are sorted by order.

TPSAInterface.setm! — Function

setm!(t, v, mono::AbstractArray{<:Integer})

Sets the coefficient of the monomial with orders mono to v.

TPSAInterface.copy! — Function

Sets the entire TPS t equal to t1, where t1 may be another TPS or a Number.

TPSAInterface.add! — Function

Sets the TPS t equal to a + b

TPSAInterface.sub! — Function

Sets the TPS t equal to a - b

TPSAInterface.mul! — Function

Sets the TPS t equal to a * b

TPSAInterface.div! — Function

Sets the TPS t equal to a / b

TPSAInterface.pow! — Function

Sets the TPS t equal to a ^ b

TPSAInterface.getord! — Function

Sets t equal to the homogenous polynomial of order ord in t1.

TPSAInterface.cutord! — Function

Sets t equal to t1 with monomials at order ord and above removed. Or, if ord is negative, will remove monomials with orders at and below abs(ord).

TPSAInterface.deriv! — Function

Sets t equal to the derivative of t1 with respect to the i-th differential.

TPSAInterface.inv! — Function

Sets the map m equal to the inverse of the map m1, ignoring any scalar part.

TPSAInterface.compose! — Function

Sets m equal to m2 ∘ m1 where m1 and/or m2 may be TPS types or arrays of TPS types.