TPSAInterface

The TPSA package extension should provide overloads for various functions.

TPSAInterface.InitGTPSA

Base.copy!

TPSAInterface.add!

TPSAInterface.clear!

TPSAInterface.compose!

TPSAInterface.cutord!

TPSAInterface.cycle!

TPSAInterface.deriv!

TPSAInterface.div!

TPSAInterface.geti

TPSAInterface.getm

TPSAInterface.getord!

TPSAInterface.imag!

TPSAInterface.inv!

TPSAInterface.is_tps

TPSAInterface.is_tps_type

TPSAInterface.maxord

TPSAInterface.mul!

TPSAInterface.ndiffs

TPSAInterface.nmonos

TPSAInterface.norm_tps

TPSAInterface.numtype

TPSAInterface.pow!

TPSAInterface.real!

TPSAInterface.scalar

TPSAInterface.seti!

TPSAInterface.setm!

TPSAInterface.sub!

TPSAInterface.InitGTPSA — Type

InitGTPSA{D,DD}

Struct used to define a TPSA using the GTPSA.jl backend.

Defined by TPSAInterface.jl.

Constructors

InitGTPSA{D,DD}(; dynamic_descriptor=nothing)

Fields

For static Descriptor resolution:

D is the Descriptor

DD is Nothing and dynamic_descriptor is nothing

For dynamic Descriptor resolution:

D == GTPSA.Dynamic

DD == Descriptor and dynamic_descriptor is said Descriptor

Base.copy! — Method

Sets the entire TPS t equal to t1, where t1 may be another TPS or a Number. Promotion is supported; e.g. if t has numtype ComplexF64, and t1 has numtype Float64), calling copy!(t, t1) is allowed

TPSAInterface.add! — Method

Sets the TPS t equal to a + b

TPSAInterface.clear! — Method

Sets all monomial coefficients in t equal to 0.

TPSAInterface.compose! — Method

Sets m equal to m2 ∘ m1 where m1 and/or m2 may be TPS types or arrays of TPS types.

TPSAInterface.cutord! — Method

Sets t equal to t1 with monomials at order ord and above removed. Or, if ord is negative, will remove monomials with orders at and below abs(ord).

TPSAInterface.cycle! — Method

Cycles through the nonzero monomial coefficients in the TPS t starting at i. Set i equal to -1 to start at the first nonzero monomial coefficient. Returns the next nonzero monomial coefficient index and will, if provided, set m_ and v_ to the monomial orders and coefficient of the next nonzero monomial coefficient, respectively.

TPSAInterface.deriv! — Method

Sets t equal to the derivative of t1 with respect to the i-th differential.

TPSAInterface.div! — Method

Sets the TPS t equal to a / b

TPSAInterface.geti — Method

geti(t, i::Integer)

Gets the i-th monomial coefficient in the TPS t, where i=0 specifies the scalar part, and the monomials are sorted by order.

TPSAInterface.getm — Method

getm(t, mono)

Gets the coefficient of the monomial with orders mono.

TPSAInterface.getord! — Method

Sets t equal to the homogenous polynomial of order ord in t1.

TPSAInterface.imag! — Method

Sets the TPS t equal to imag(t1)

TPSAInterface.inv! — Method

Sets the map m equal to the inverse of the map m1, ignoring any scalar part.

TPSAInterface.is_tps — Method

Returns TPSAInterface.IsTPS() iftis a TPS, elseTPSAInterface.IsNotTPS().

TPSAInterface.is_tps_type — Method

Returns TPSAInterface.IsTPSType() iftis a TPS type, elseTPSAInterface.IsNotTPSType().

TPSAInterface.maxord — Method

Returns the maximum truncation order of the TPSA. t may be a TPS, TPS Type, or a AbstractTPSAInit.

TPSAInterface.mul! — Method

Sets the TPS t equal to a * b

TPSAInterface.ndiffs — Method

Returns the number of variables + parameters in the TPSA. t may be a TPS, TPS Type, or a AbstractTPSAInit.

TPSAInterface.nmonos — Method

Returns the number of monomial coefficients in the TPSA. t may be a TPS, TPS Type, or a AbstractTPSAInit.

TPSAInterface.norm_tps — Method

Returns a norm of the passed TPS t including all monomial coefficients.

TPSAInterface.numtype — Method

Returns the number type of the monomial coefficients of the TPS. t may be a TPS or TPS Type.

TPSAInterface.pow! — Method

Sets the TPS t equal to a ^ b

TPSAInterface.real! — Method

Sets the TPS t equal to real(t1)

TPSAInterface.scalar — Method

Returns the zeroth order (scalar) part of the TPS t.

TPSAInterface.seti! — Method

seti!(t, v, i::Integer)

Sets the i-th monomial coefficient in the TPS t to v, where i=0 specifies the scalar part, and the monomials are sorted by order.

TPSAInterface.setm! — Method

setm!(t, v, mono)

Sets the coefficient of the monomial with orders mono to v.

TPSAInterface.sub! — Method

Sets the TPS t equal to a - b